move the business logic outside
